requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.9.1.post1
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==0.28.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
requests==2.32.3
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
yarl==1.18.3
